demo(m14-g8): Steps 2–5 Demo 5 prep — ZMB screenshot brief, narrated spec, demo.sh, walkthroughs
- docs/demo/m14/screenshot-brief.md: five-frame brief (UX Designer Agent)
  - Thesis frame C first: Grounding strip citation answering reserve challenge
  - Frames A/B/D/E: Grounding strip at load, Zone 1B Layer 3, PSP, evidence thread
  - Single entity ZMB; no Mode 3; presentation order C→A→B→D→E
- frontend/tests/e2e/demo-narrated.spec.ts: rewritten for M14 Demo 5
  - ZMB ECF scenario via API; political_economy module enabled
  - Six-step trajectory; IMF program acceptance step 1; fiscal conditionality step 2
  - NM-039 sentinel present; 1440×900 viewport (Issue #675)
  - Screenshots to docs/demo/m14/screenshots/
- frontend/tests/e2e/demo-narrated-m12.spec.ts: M12 archive (Jordan/Egypt Demo 4)
- scripts/demo.sh: updated for M14 Demo 5 ZMB/no-Mode3
  - Presenter guide: challenge-response framing; five-step sequence C→A→B→D→E
  - Honest disclosures updated: ZMB not yet backtested; PSP exploratory; no Mode 3
  - NM-041 gate: bash -n exits 0 (verified)
- docs/demo/stakeholder-walkthrough.md: root pointer updated to v0.14.0
